Remove methods chapter and user requirements analysis
The requirements were not really referred to in the thesis later on and therefore stood alone. The separate methods chapter felt weird in the way that methodology is described in 3 different locations: introduction, methods chapter and the different other chapters. I will try to incorporate everything that used to be in the methods chapter into the other chapters, for now I did that for the second usability study and the user task analysis.